Fix MSVC warning about file naming collision
